refactor(lib): move string helpers to dedicated util file
