+++ apps/manage/src/app/views/cases/create-a-case/question-utils.js
@@ -2,8 +2,7 @@ import AddressValidator from '@planning-inspectorate/dynamic-forms/src/validator
 import RequiredValidator from '@planning-inspectorate/dynamic-forms/src/validator/required-validator.js';
 import StringValidator from '@planning-inspectorate/dynamic-forms/src/validator/string-validator.js';
 import { COMPONENT_TYPES } from '@planning-inspectorate/dynamic-forms';
-import { camelCaseToUrlCase } from '@pins/crowndev-lib/util/string.js';
-import { capitalize } from '@planning-inspectorate/dynamic-forms/src/lib/utils.js';
+import { camelCaseToUrlCase, sentenceCase } from '@pins/crowndev-lib/util/string.js';
 import MultiFieldInputValidator from '@pins/crowndev-lib/validators/multi-field-input-validator.js';
 import TelephoneNumberValidator from '@pins/crowndev-lib/validators/telephone-number-validator.js';
 import EmailValidator from '@planning-inspectorate/dynamic-forms/src/validator/email-validator.js';
@@ -114,7 +113,7 @@ export function multiContactQuestions({ prefix, title, options }) {
 	const questions = {};
 	questions[`${prefix}ContactDetails`] = {
 		type: CUSTOM_COMPONENTS.CUSTOM_MULTI_FIELD_INPUT,
-		title: `${capitalize(title)} contact`,
+		title: `${sentenceCase(title)} contact`,
 		question: `Add ${title} contact details`,
 		fieldName: `${prefix}ContactDetails`,
 		url: `${prefixUrl}-contact`,
