+++ apps/manage/src/app/views/cases/view/question-utils.test.js
@@ -1,12 +1,6 @@
 import { describe, it } from 'node:test';
 import assert from 'node:assert';
-import {
-	contactQuestions,
-	dateQuestion,
-	eventQuestions,
-	filteredStagesToRadioOptions,
-	camelCaseToSentenceCase
-} from './question-utils.js';
+import { contactQuestions, dateQuestion, eventQuestions, filteredStagesToRadioOptions } from './question-utils.js';
 import { APPLICATION_PROCEDURE_ID, APPLICATION_STAGE_ID } from '@pins/crowndev-database/src/seed/data-static.js';
 
 describe('question-utils', () => {
@@ -255,14 +249,6 @@ describe('question-utils', () => {
 				{ id: APPLICATION_STAGE_ID.DECISION, displayName: 'Final decision' }
 			];
 			assert.deepStrictEqual(result, expected);
-		});
-	});
-	describe('camelCaseToSentenceCase', () => {
-		it('should turn a basic camelCaseSentence into Sentence case', () => {
-			const sentence = camelCaseToSentenceCase('thisIsAnExample');
-
-			assert.ok(sentence);
-			assert.strictEqual(sentence, 'This is an example');
 		});
 	});
 });
