+++ packages/lib/util/string.test.js
@@ -1,6 +1,6 @@
 import { describe, it } from 'node:test';
 import assert from 'node:assert';
-import { camelCaseToUrlCase } from './string.js';
+import { camelCaseToUrlCase, camelCaseToSentenceCase, sentenceCase } from './string.js';
 
 describe('string util', () => {
 	describe('camelCaseToUrlCase', () => {
@@ -27,6 +27,28 @@ describe('string util', () => {
 
 		it('handles empty string input', () => {
 			assert.strictEqual(camelCaseToUrlCase(''), '');
+		});
+	});
+	describe('camelCaseToSentenceCase', () => {
+		it('should turn a basic camelCaseSentence into Sentence case', () => {
+			const sentence = camelCaseToSentenceCase('thisIsAnExample');
+
+			assert.ok(sentence);
+			assert.strictEqual(sentence, 'This is an example');
+		});
+		it('should handle consecutive capital letters correctly', () => {
+			const sentence = camelCaseToSentenceCase('thisIsHTMLParser');
+
+			assert.ok(sentence);
+			assert.strictEqual(sentence, 'This is h t m l parser');
+		});
+	});
+	describe('sentenceCase', () => {
+		it('should turn a string into sentence case', () => {
+			const sentence = sentenceCase('string with a Proper Noun');
+
+			assert.ok(sentence);
+			assert.strictEqual(sentence, 'String with a Proper Noun');
 		});
 	});
 });
